Progress towards CUnit on IOP.
* Change tack - now making object files and a separate .a library.
  Ultimately I'd like to compile the object files for ps2-cunit-port
  *into* libcunit.a, but this will do for the time being.

* Add some missing standard library methods using the usual import
syntax. I don't think it's doing to be possible to put the import tables
into a .a, so it's there as documentation too - just needs adding to
the README I guess.

Basically all that's left afaict is floating point arithmetic (I'm
probably gonna go with Berkeley SoftFloat) and the clock function which
I think I'm gonna have to make myself. yikes.